sync, valgrind campain fixes, and other small fixes

+++ tbx/byar.h
@@ -14,8 +14,8 @@ pbyar_t byar_destroy(pbyar_t b);
 pbyar_t byar_new_from_data(char *data, size_t len);
 
 int     byar_data_get(pbyar_t b, char **data, size_t *size);
-int   	byar_push(pbyar_t b, int len, char *data);
-char *	byar_pop(pbyar_t b, int len, char *data);
+int   	byar_push(pbyar_t b, size_t len, char *data);
+char *	byar_pop(pbyar_t b, size_t len, char *data);
 int     byar_rewind(pbyar_t b);
 
 #ifdef __cplusplus
